Added Blur to Overlays and Fixed Tools without Logos
Added a blur class to add a blur, added a filter-transition class for
things that will have changing CSS filters, and made showing or hiding
an overlay toggle the blur class on the header, footer, and
page-container, which all now have the filter-transition class to
animate the blur being added and removed. Also added functionality to
fix C# and Google Charts tool display, as they do not have logos.